+++ server/tsuserver.py
@@ -47,8 +47,8 @@ class TsuserverDR:
         self.release = 4
         self.major_version = 2
         self.minor_version = 5
-        self.segment_version = ''
-        self.internal_version = '200731a'
+        self.segment_version = 'post1'
+        self.internal_version = '200803a'
         version_string = self.get_version_string()
         self.software = 'TsuserverDR {}'.format(version_string)
         self.version = 'TsuserverDR {} ({})'.format(version_string, self.internal_version)